Added a Main Menu and a new Unblocking Function (no specific times involved)

+++ Blocker.py
@@ -3,18 +3,23 @@ from datetime import datetime
 
 # The path to your Windows hosts file
 hosts_path = r"C:\Windows\System32\drivers\etc\hosts"
-
-# Where we redirect blocked sites (your own computer)
 redirect = "127.0.0.1"
-
-#Websites blocked
 website_list = ["www.facebook.com", "facebook.com"]
 
-while True:
-    #Check current time
-    if 9 <= datetime.now().hour < 14:
-        print("Blocking Websites")
+def MainMenu():
+    print("------------------------------------------")
+    print("WELCOME TO WEBSITE BLOCKER")
+    print("------------------------------------------\n" )
+    time.sleep(1)
+    print("1. Start Blocking")
+    print("2. Exit")
+    
+    choice = input("What do you want to do?\n")
+    if choice == "1":
+        Block()
 
+def Block():
+    try:
         with open(hosts_path, 'r+') as file:
             content = file.read()
             for website in website_list:
@@ -22,15 +27,27 @@ while True:
                     pass
                 else:
                     file.write(redirect + " " + website + "\n")
-    else:
-        print("Free time! Unblocking Websites")
+        
+        print("\nWebsites Blocked")    
+        
+        stop = input("Unblock the websites by typing stop: ").lower() 
+        
+        if stop == "stop":
+            with open(hosts_path, "r+") as file:
+                content = file.readlines()
+                file.seek(0)
+                for line in content:
+                    if not any(website in line for website in website_list):
+                        file.write(line)
+                file.truncate()
+            
+            print("Websites Unblocked")
+            time.sleep(2)
+            MainMenu()
+            
+    except PermissionError:
+        print("\n[ERROR]: You must run your terminal/editor as ADMINISTRATOR for it to work.")
+        time.sleep(3)
+        MainMenu()
 
-        with open(hosts_path, "r+") as file:
-            content = file.readlines()
-
-            file.seek(0)
-            for line in content:
-                if not any(website in line for website in website_list):
-                    file.write(line)
-            file.truncate()
-    time.sleep(300)
+MainMenu()
